pyproject.toml
[project]
name = "geminikeymanagement"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-api-keys>=0.5.17",
    "google-cloud-resource-manager>=1.14.2",
    "google-cloud-service-usage>=1.13.1",
    "jsonschema>=4.25.1",
    "colorama>=0.4.6"
]
